experimental PI controller (change integrator)
git-svn-id: https://openhr20.svn.sourceforge.net/svnroot/openhr20@295 e6282f77-c345-0410-9e5f-9d7f1dee61d8

+++ rfmsrc/OpenHR20/controller.h
@@ -64,5 +64,5 @@ void CTL_change_mode(int8_t dif);
 #define CTL_ERR_NA_1                    (1<<1)
 #define CTL_ERR_NA_0                    (1<<0)
 
-extern int16_t sumError;
+extern int32_t sumError;
